+++ src/modes/fs/mod.rs
@@ -2,8 +2,6 @@ mod fileinfo;
 mod icon;
 mod users;
 
-pub use fileinfo::{
-    extract_datetime, extract_extension, is_not_hidden, permission_mode_to_str, FileInfo, FileKind,
-};
+pub use fileinfo::{extract_datetime, extract_extension, is_not_hidden, FileInfo, FileKind};
 pub use icon::*;
 pub use users::Users;
