+++ src/modes/menu/mod.rs
@@ -45,7 +45,7 @@ pub use mount::*;
 pub use node_creation::NodeCreation;
 pub use nvim::nvim;
 pub use password::{PasswordHolder, PasswordKind, PasswordUsage};
-pub use permissions::{parse_input_permission, Permissions, MAX_FILE_MODE};
+pub use permissions::{parse_input_permission, permission_mode_to_str, Permissions, MAX_FILE_MODE};
 pub use picker::*;
 pub use regex::{regex_flagger, CaseDependantRegex};
 pub use remote::Remote;
